pyproject.toml
[project]
name = "mlbio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kagglehub>=0.4.0",
    "lime>=0.2.0.1",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "scikit-image>=0.25.0",
    "scikit-learn>=1.8.0",
    "streamlit>=1.41.0",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]